moved old Pathfinder to pathfinder.old package
