+++ core/src/main/java/com/strategy_bit/chaos_brawl/ashley/components/MovementComponent.java
@@ -4,9 +4,6 @@ import com.badlogic.ashley.core.Component;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Queue;
-import com.strategy_bit.chaos_brawl.ashley.systems.MovementSystem;
-import com.strategy_bit.chaos_brawl.pathfinder.Pathfinder;
-import com.strategy_bit.chaos_brawl.world.World;
 
 /**
  * component that has fields for moving an entity
@@ -17,7 +14,8 @@ import com.strategy_bit.chaos_brawl.world.World;
  */
 
 public class MovementComponent implements Component {
-
+    //TODO maybe remove targetLocation and replace it completely with path
+    //TODO give movementComponent a field that tells him to stop move but twhich that will be set from different systems
     /**
      * Will be changed over time
      */
