+++ core/src/main/java/com/strategy_bit/chaos_brawl/ashley/systems/MovementSystem.java
@@ -7,12 +7,10 @@ import com.badlogic.ashley.core.Family;
 import com.badlogic.ashley.systems.IteratingSystem;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.utils.Array;
 import com.strategy_bit.chaos_brawl.ashley.components.MovementComponent;
 import com.strategy_bit.chaos_brawl.ashley.components.CombatComponent;
 import com.strategy_bit.chaos_brawl.ashley.components.TransformComponent;
 import com.strategy_bit.chaos_brawl.util.VectorMath;
-import com.strategy_bit.chaos_brawl.pathfinder.Pathfinder;
 
 
 /**
@@ -44,6 +42,7 @@ public class MovementSystem extends IteratingSystem {
 
     @Override
     protected void processEntity(Entity entity, float deltaTime) {
+        //TODO moved this code part to somewhere else
         CombatComponent combatComponent=mCombatComponent.get(entity);
         if (combatComponent!=null){
             //Unit is attacking
@@ -66,7 +65,7 @@ public class MovementSystem extends IteratingSystem {
         // velocity = normalVector(targetLocation - position) * speed
         Vector2 velocity = VectorMath.scl(VectorMath.nor(VectorMath.sub(targetLocation, position)), speed);
         movementComponent.setVelocity(velocity);
-
+        //TODO Florian rotate entities here try to find out the direction of velocity
         // position = position + (velocity * deltaTime)
         transform.setPosition(VectorMath.add(position, VectorMath.scl(movementComponent.getVelocity(), Gdx.graphics.getDeltaTime())));
