+++ core/src/main/java/com/strategy_bit/chaos_brawl/pathfinder/old/Node.java
@@ -1,4 +1,4 @@
-package com.strategy_bit.chaos_brawl.pathfinder;
+package com.strategy_bit.chaos_brawl.pathfinder.old;
 
 import com.badlogic.gdx.math.Vector2;
